requirements.txt
astroid==2.4.1
fpdf2==2.0.5
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.5
Pillow==6.2.2
pylint==2.5.2
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
